bug fixes (#9)
* Convert milliseconds to deciseconds to match spec

rfc8010 indicates that the dateTime type should be:

OCTET-STRING consisting of eleven octets whose contents are defined by
"DateAndTime" in RFC 2579 [RFC2579]
https://tools.ietf.org/html/rfc8010#page-25

DateAndTime from rfc2579:
https://tools.ietf.org/html/rfc2579#page-18

* Increment month when deserializing dateTime

getMonth() is 0 based

* Write dateTime sign to buffer as single char

* Install roboter-cli locally

so it doesn't need be installed globally

* Allow passing an ordered array of groups

Some ipp operations such as get-jobs requires serializing more than one
of the same attribute group tag.  The previous model didn't allow
sending multiple as the attribute tag was the key to the object which is
passed to the serializer.  The serializer will now check if the item in
the object is an array and then add each item of the array, preserving
order.